[Refactor]: RedisTokenRepository->TokenRepository
