+++ src/main/java/io/dev/concertreservationsystem/common/config/redis/RedisCacheConfig.java
@@ -43,8 +43,8 @@ public class RedisCacheConfig {
      * @return RedisTemplate<String, String>
      */
     @Bean
-    public RedisTemplate<String, String> redisTemplate() {
-        RedisTemplate<String, String> redisTemplate = new RedisTemplate<>();
+    public RedisTemplate<String, Object> redisTemplate() {
+        RedisTemplate<String, Object> redisTemplate = new RedisTemplate<>();
         redisTemplate.setKeySerializer(new StringRedisSerializer());
         redisTemplate.setValueSerializer(new GenericJackson2JsonRedisSerializer());
         redisTemplate.setConnectionFactory(defaultRedisConnectionFactory());
