+++ src/main/java/io/dev/concertreservationsystem/domain/token/redis/RedisToken.java
@@ -5,10 +5,12 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.io.Serializable;
+
 @Getter
 @Setter
 @Builder
-public class RedisToken {
+public class RedisToken implements Serializable {
     private String tokenId;
     private String userId;
