+++ src/main/java/io/dev/concertreservationsystem/infrastructure/token/RedisTokenRepositoryImpl.java
@@ -1,21 +1,19 @@
 package io.dev.concertreservationsystem.infrastructure.token;
 
 import io.dev.concertreservationsystem.common.config.redis.CacheKey;
-import io.dev.concertreservationsystem.domain.token.Token;
-import io.dev.concertreservationsystem.domain.token.redis.RedisTokenRepository;
+import io.dev.concertreservationsystem.domain.token.redis.TokenRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.ZSetOperations;
 import org.springframework.stereotype.Repository;
 
-import java.util.Objects;
 import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 
 @Repository
 @RequiredArgsConstructor
-public class RedisTokenRepositoryImpl implements RedisTokenRepository {
+public class RedisTokenRepositoryImpl implements TokenRepository {
 
     private final RedisTemplate<String, String> redisTemplate;
